Successfuly added sprite #0 hit detection.
Added sprite #0 hit detection with a 1 tick delay. I've also changed the
branch instructions to remove the pointer reference and instead use the
same CPU structs as the other opcode functions. This change meant that
the CPU execution function no longer needed a pointer to the PC, which
is replaced by read_addr() helper function. Minor changes have been made
to the other files such as comment adjusments. I also removed the
fetch_16_IND() function as it was pointless and not implemented.